Add a method for factorial iterative
Created a number variable equal to 1 so that it can start counting
upward from 1 to the amount of number the input is. The if statement is
used to return the 1 if there is a 0 input. The for loop is used
to keep track of the amount of iteration it goes through.