Add shuffle method to file
The shuffle method takes in an array and does two things, 1) it checks
if it has any symbols, if so, it will return an unshuffl array. 2)
if it passes the first condition, the second part of the body will loop
through the amount of element in the array and swap the last index of the
element with a random element generated from 0 to one less the last element.